+++ packages/member-apps/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [2.1.1](https://bitbucket.org/openagenda/oa/branches/compare/@openagenda/member-apps@2.1.1..@openagenda/member-apps@2.1.0) (2020-11-23)
+
+**Note:** Version bump only for package @openagenda/member-apps
+
+
+
+
+
 ## [2.1.0](https://bitbucket.org/openagenda/oa/branches/compare/@openagenda/member-apps@2.1.0..@openagenda/member-apps@2.0.13) (2020-11-23)