+++ packages/supervisor/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [1.3.1](https://bitbucket.org/openagenda/oa/branches/compare/@openagenda/supervisor@1.3.1..@openagenda/supervisor@1.3.0) (2020-11-23)
+
+**Note:** Version bump only for package @openagenda/supervisor
+
+
+
+
+
 ## [1.3.0](https://bitbucket.org/openagenda/oa/branches/compare/@openagenda/supervisor@1.3.0..@openagenda/supervisor@1.2.6) (2020-11-23)